Feat: DivisorProduct (fixed)

+++ src/Loops/Exercise019_DivisorProduct.java
@@ -6,9 +6,9 @@ public class Exercise019_DivisorProduct {
     static void main(String[] args) {
             Scanner sc = new Scanner(System.in);
 
-            int n = sc.nextInt();
+            long n = sc.nextLong();
 
-            int sumProduct = 1;
+            long sumProduct = 1;
 
             for(int i = 1; i <= n; i++){
                 if (n % i == 0){
